+++ apps/ple_demo_base/lib/ple_demo_base_web/live/test2_live.ex
@@ -1,7 +1,9 @@
 defmodule PleDemoBaseWeb.Test2Live do
 
   use PleDemoBaseWeb, :live_view
-  use Phoenix.LiveEditable
+  use Phoenix.LiveEditableView
+
+  import PleDemoBase.Components
 
   # ----- lifecycle callbacks -----
 
@@ -9,11 +11,8 @@ defmodule PleDemoBaseWeb.Test2Live do
     {:ok, assign(socket, id: "IdOne", data: "Click Me to Edit")}
   end
 
-  # ----- event handlers -----
-
-  def handle_event("save", %{"data" => data}, socket) do
-    new_socket = assign(socket, data: data, focusid: "NONE")
-    {:noreply, new_socket}
+  def handle_params(_params, _url, socket) do
+    {:noreply, socket}
   end
 
   # ----- helpers -----