+++ apps/ple_demo_base/lib/ple_demo_base_web/router.ex
@@ -18,7 +18,11 @@ defmodule PleDemoBaseWeb.Router do
     pipe_through :browser
 
     live "/", PageLive
+    live "/home", PageLive
     live "/test1", Test1Live
     live "/test2", Test2Live
+    live "/test3", Test3Live
+    live "/test4", Test4Live
+    live "/test5", Test5Live
   end
 end