+++ apps/ple_util/test/ple_util_test.exs
@@ -1,0 +1,8 @@
+defmodule PleUtilTest do
+  use ExUnit.Case
+  doctest PleUtil
+
+  test "greets the world" do
+    assert PleUtil.hello() == :world
+  end
+end