+++ apps/phoenix_live_editable/CHANGELOG.md
@@ -3,3 +3,9 @@
 ## 0.0.1-alpha.1 (2022-Jun-03)
 
 - Initial Checkin
+
+## 0.0.1-alpha.2 (2022-Jun-15)
+
+- Settled on basic code organization (view and component layers) 
+- Created the ViewCache to store configuration data
+- Created interfaces for Milligram and Tailwind3